requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
google-generativeai

python-dotenv
fake-useragent

thefuzz
pydantic
fpdf2
openpyxl
pypdf
python-docx
python-multipart
sqlalchemy
asyncpg
greenlet
